Remove booking functionality moved into the view bookings page so you can select and delete an entry
